Update `UIColor` initialisation using `hex`
