--- Source/Extensions/UIColor+Extension.swift
@@ -1,33 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Ehab Alsharkawy on 27/04/2022.
-//
-
-import Foundation
-import UIKit
-
-extension UIColor {
-    convenience init(hex: String, alpha: CGFloat = 1.0) {
-        var cString = hex.trimmingCharacters(in: .whitespacesAndNewlines).uppercased()
-
-        if cString.hasPrefix("#") { cString.removeFirst() }
-
-        guard cString.count == 6 else {
-            self.init(hex: "ff0000") // gray color
-            return
-        }
-
-        var rgbValue: UInt64 = 0
-        Scanner(string: cString).scanHexInt64(&rgbValue)
-
-        self.init (
-            red: CGFloat((rgbValue & 0xFF0000) >> 16) / 255.0,
-            green: CGFloat((rgbValue & 0x00FF00) >> 8) / 255.0,
-            blue: CGFloat(rgbValue & 0x0000FF) / 255.0,
-            alpha: CGFloat(1.0)
-        )
-    }
-
-}
